+++ tests/Pest.php
@@ -1,6 +1,6 @@
 <?php
 
-use Mint\Translatable\Tests\TestCase;
+use Rat\Translatable\Tests\TestCase;
 
 if (!function_exists('get_post_class')) {
     function get_post_class() {