+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-mint/laravel-translatable / Changelog
-=====================================
+rat.md/laravel-translatable / Changelog
+=======================================
+
+## Version 0.1.1 (Beta)
+- Change Composer namespace.
 
 ## Version 0.1.0 (Beta)
 - Initial Release.